Provider is being get for each checkpoint
In checkpoints/views.py, IndexView class, get_data() method, we list
the checkpoints for a provider.
After that, we iterate the checkpoints for that provider and get the
provider using karbor client. This is not necessary: all the checkpoints
are from the same provider, and even if they weren't, we could use karbor
client to get each provider only once.

Change-Id: I790116137a321570c6d386e435f24a5efde2e607
Closes-Bug: #1643331