Fix checkpoints pagination error
Currently, previous page in checkpoints pagination not
work, it always throw an error in the web page. It is
because that checkpoints in karbor do not need reverse
the order (checkpoints are sorted by 'timestamp@id' in
the backend, so we can not do the reverse only be the
id or name)

This patch will fix it.

Change-Id: I45c3e7337bf2ee9d3f2fbf9ee7a438884b8b8ca9
Closes-Bug: #1714909