translation: drop babel extractor definitions
babel extractors are now registered via python entry points,
so there is no need to declare babel extractors in babel configs.

This change is important to make translation work in Django 2.2.
django-babel does not work with Django 2.2 and looks unmaintained
for over two years. The horizon team is thinking to switch the extractor
to enmerkar (a fork of django-babel) to make extraction of translation
string work again near future. It is important to drop the extractor
definition to make the transition smooth.

django-babel (and enmerkar, a successor of django-babel) will be
installed via horizon, so we don't need to have it in requirements.txt
explicitly. It would make the transition of the babel extractor easy.
Babel is dropped from requirements.txt from the same reason.

Django dependency is also upadted to sync global-requirements.txt
to pass the requirements-check job.

Change-Id: I093d2c9789a142e909da6f3bedc12a91b181ab02