Fix horizon dependency
This commit moves horizon to requirements.txt.

horizon was added to uper-constraints in stable branches recently,
so installing from the tarball no longer works. upper-constraints
in stable branches picks up the right version of horizon, so there
is no need for the upper bound of the horizon version.

In addition, the link of the horizon tarball was not updated after
the stable branch was created. I don't know why it works for stable/train
but I believe this commit mitigates the situation as the upper-constraints
file in the requirements repo picks up the right version of horizon.

This commit is proposed directly as the similar change in the master
branch happened as part of Django 1.11 drop.

Change-Id: I6e2d30e91373e1cc742a4f3c02d32a90b714d7e4